wip: move the qr scan to issuers and verifiers pages
