+++ src/pages/Verifiers/Verifiers.js
@@ -1,9 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { FaShare } from 'react-icons/fa';
-
+import {BsQrCodeScan} from 'react-icons/bs'
 import * as api from '../../api';
 import Layout from '../../components/Layout';
 import Spinner from '../../components/Spinner';
+import QRCodeScanner from '../../components/QRCodeScanner'; // Replace with the actual import path
 
 function highlightBestSequence(verifier, search) {
   if (typeof verifier !== 'string' || typeof search !== 'string') {
@@ -97,10 +98,34 @@ const Verifiers = () => {
 			});
 	};
 
+	// qr code part
+	const [isQRScannerOpen, setQRScannerOpen] = useState(false);
+
+	const openQRScanner = () => {
+		setQRScannerOpen(true);
+	};
+
+	const closeQRScanner = () => {
+		setQRScannerOpen(false);
+	};
 
   return (
     <Layout>
       <div className="px-4 sm:px-6">
+				<div className="flex justify-between items-center">
+          <h1 className="text-2xl font-bold text-custom-blue">Send Credentials</h1>
+          <button
+            className="px-2 py-2 mb-2 text-white bg-custom-blue hover:bg-custom-blue-hover focus:ring-4 focus:outline-none focus:ring-custom-blue font-medium rounded-lg text-sm px-4 py-2 text-center dark:bg-custom-blue-hover dark:hover:bg-custom-blue-hover dark:focus:ring-custom-blue-hover"
+            onClick={openQRScanner} // Open the QR code scanner modal
+          >
+            <div className="flex items-center">
+              <BsQrCodeScan size={20} className="text-white mr-2 sm:inline" />
+							<span className="sm:inline">Scan</span>
+
+            </div>
+          </button>
+          
+        </div>
         <h1 className="text-2xl mb-2 font-bold text-custom-blue">Verifiers</h1>
         <hr className="mb-2 border-t border-custom-blue/80" />
         <p className="italic text-gray-700">Search and choose a verifier for credential retrieval</p>
@@ -136,8 +161,6 @@ const Verifiers = () => {
         )}
       </div>
 
-
-			
 			{showPopup && (
 				<div className="fixed inset-0 flex items-center justify-center z-50">
 					<div className="absolute inset-0 bg-black opacity-50"></div>
@@ -191,6 +214,13 @@ const Verifiers = () => {
 				</div>
 			)}
 
+			{/* QR Code Scanner Modal */}
+			{isQRScannerOpen && (
+        <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-75">
+          {/* Include your QR code scanner component here */}
+          <QRCodeScanner onClose={closeQRScanner} /> {/* Pass a prop to close the modal */}
+        </div>
+      )}
     </Layout>
   );
 };
